fix: address code review feedback for customer selector component
- Fix avatar image overlay to only show on error (not always covering the image)
- Add customerId parameter to all onSelect callbacks for proper contract
- Extract formatDate to shared registry/lib/date.ts to avoid duplication
- Add 'use client' directive to registry-index.ts to fix RSC violation
