+++ platform/flowglad-next/src/registry/base/invoices-list/utils.ts
@@ -1,21 +1,4 @@
-export function formatDate(
-  date: Date | string,
-  locale: string = 'en-US',
-  options?: Intl.DateTimeFormatOptions
-): string {
-  const dateObj = typeof date === 'string' ? new Date(date) : date
-
-  const defaultOptions: Intl.DateTimeFormatOptions = {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-    ...options,
-  }
-
-  return new Intl.DateTimeFormat(locale, defaultOptions).format(
-    dateObj
-  )
-}
+export { formatDate } from '@/registry/lib/date'
 
 export function getInvoiceStatusColor(status: string): string {
   const statusColors: Record<string, string> = {
