+++ platform/flowglad-next/src/registry/lib/date.ts
@@ -1,0 +1,18 @@
+export function formatDate(
+  date: Date | string,
+  locale: string = 'en-US',
+  options?: Intl.DateTimeFormatOptions
+): string {
+  const dateObj = typeof date === 'string' ? new Date(date) : date
+
+  const defaultOptions: Intl.DateTimeFormatOptions = {
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+    ...options,
+  }
+
+  return new Intl.DateTimeFormat(locale, defaultOptions).format(
+    dateObj
+  )
+}
